livedata for settings and list of songs
