+++ app/src/main/AndroidManifest.xml
@@ -20,10 +20,6 @@
         android:theme="@style/Theme.AppCompat.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".settings.Settings"
-            android:parentActivityName=".MainActivity"
-            android:exported="false" />
-        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
