+++ app/src/main/java/com/sirhadrian/musicplayer/ui/SongsListFragment.java
@@ -1,6 +1,5 @@
 package com.sirhadrian.musicplayer.ui;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -13,6 +12,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -21,7 +22,7 @@ import androidx.viewpager2.widget.ViewPager2;
 import com.sirhadrian.musicplayer.R;
 import com.sirhadrian.musicplayer.databinding.FragmentSongsListBinding;
 import com.sirhadrian.musicplayer.model.SongModel;
-import com.sirhadrian.musicplayer.settings.Settings;
+import com.sirhadrian.musicplayer.settings.SettingsFragment;
 import com.sirhadrian.musicplayer.settings.SettingsViewModel;
 import com.sirhadrian.musicplayer.utils.Result;
 import com.sirhadrian.musicplayer.utils.ResultCallback;
@@ -38,9 +39,10 @@ public class SongsListFragment extends Fragment {
 
     private SharedDataViewModel mSharedData;
     private SettingsViewModel mSettings;
+    private SongsListViewModel mSongsListObserved;
 
     private RecyclerView mRecyclerView;
-    private SongsAdapter mSongsAdapter;
+    private SongsAdapter mSongsAdapter; // update in viewmodel with recyclerview
     private final ViewPager2 viewPager2Activity;
 
     private String searchFolder;
@@ -61,60 +63,22 @@ public class SongsListFragment extends Fragment {
         View view = binding.getRoot();
 
         mSharedData = new ViewModelProvider(requireActivity()).get(SharedDataViewModel.class);
+        mSongsListObserved = new ViewModelProvider(requireActivity()).get(SongsListViewModel.class);
+        mSongsListObserved.get_mSongsList().observe(getViewLifecycleOwner(), new Observer<List<SongModel>>() {
+            @Override
+            public void onChanged(List<SongModel> songModels) {
+                mSongsAdapter = new SongsAdapter(songModels);
+                mRecyclerView.setAdapter(mSongsAdapter);
+            }
+        });
+
         mSettings = new ViewModelProvider(requireActivity()).get(SettingsViewModel.class);
         mSettings.get_mDirPath().observe(getViewLifecycleOwner(), s -> searchFolder = s);
 
         mRecyclerView = binding.fragmentSongsListRecyclerView;
         mRecyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
 
-        Toolbar toolbar = requireActivity().findViewById(R.id.my_toolbar);
-        toolbar.setOnMenuItemClickListener(item -> {
-                    switch (item.getItemId()) {
-                        case R.id.settings:
-                            makeSettingsRequest();
-                            break;
-
-                        case R.id.scan:
-                            makeScanRequest(searchFolder, result -> {
-                                if (result instanceof Result.Success) {
-                                    mSongsList = ((Result.Success<List<SongModel>>) result).get_Data();
-                                    if (mSongsList != null) {
-                                        mSongsAdapter = new SongsAdapter(mSongsList);
-                                        mRecyclerView.setAdapter(mSongsAdapter);
-                                    }
-                                } else if (result instanceof Result.Error) {
-                                    ((Result.Error<List<SongModel>>) result).exception.printStackTrace();
-                                }
-                            });
-                            Log.d("scan", "End Scan");
-                            break;
-
-                        default:
-                            return false;
-                    }
-
-                    return true;
-                }
-        );
         return view;
-    }
-
-    private void makeSettingsRequest() {
-        Intent intent = new Intent(getContext(), Settings.class);
-        startActivity(intent);
-    }
-
-    private void makeScanRequest(String folder, ResultCallback<List<SongModel>> callback) {
-        ExecutorService executor = Executors.newSingleThreadExecutor();
-        executor.execute(() -> {
-            try {
-                Result<List<SongModel>> result = new Result.Success<>(getPlayList(folder));
-                callback.onComplete(result);
-            } catch (Exception e) {
-                Result<List<SongModel>> errorResult = new Result.Error<>(e);
-                callback.onComplete(errorResult);
-            }
-        });
     }
 
     private class SongsAdapter extends RecyclerView.Adapter<SongsAdapter.SongHolder> {
@@ -179,30 +143,5 @@ public class SongsListFragment extends Fragment {
         }
     }
 
-    List<SongModel> getPlayList(String rootPath) {
-        List<SongModel> fileList = new ArrayList<>();
 
-        try {
-            File rootFolder = new File(rootPath);
-            File[] files = rootFolder.listFiles();
-            assert files != null;
-            for (File file : files) {
-                if (file.isDirectory()) {
-                    if (getPlayList(file.getAbsolutePath()) != null) {
-                        fileList.addAll(getPlayList(file.getAbsolutePath()));
-                    } else {
-                        break;
-                    }
-                } else if (file.getName().endsWith(".mp3")) {
-                    SongModel song = new SongModel(file.getName(), file.getAbsolutePath());
-                    fileList.add(song);
-                }
-            }
-
-            return fileList;
-        } catch (Exception e) {
-            Log.d("scan", e.toString());
-        }
-        return null;
-    }
 }
